refactor: QR 기능 관련 인터페이스 위치를 도메인 서비스로 변경
