+++ qrust-domain/src/main/java/com/qrust/qrcode/infrastructure/zxing/QrCodeDecoderImpl.java
@@ -7,7 +7,7 @@ import com.google.zxing.MultiFormatReader;
 import com.google.zxing.Result;
 import com.google.zxing.client.j2se.BufferedImageLuminanceSource;
 import com.google.zxing.common.HybridBinarizer;
-import com.qrust.qrcode.application.QrCodeDecoder;
+import com.qrust.qrcode.domain.service.QrCodeDecoder;
 import com.qrust.utils.AesUtil;
 import com.qrust.qrcode.domain.entity.vo.QrCodeData;
 import java.awt.image.BufferedImage;
